requirements.txt
asgiref==3.8.1
Brotli==1.1.0
dj-database-url==2.1.0
Django==4.2.13
djangorestframework==3.15.1
gunicorn==22.0.0
packaging==24.0
psycopg2-binary==2.9.9
sqlparse==0.5.0
typing_extensions==4.11.0
whitenoise==6.6
django-cors-headers==4.3.1
python-dotenv==1.0.0
urllib3==2.0.3
python-decouple
channels==4.1.0

# certifi==2023.5.7
# cffi==1.15.1

# charset-normalizer==3.1.0
# click==8.1.3
# cryptography==41.0.1
# h11==0.14.0
# idna==3.4
# isodate==0.6.1
# msal==1.22.0
# msal-extensions==1.0.0
# pandas==2.0.1
# portalocker==2.7.0
# pycparser==2.21
# 
# python-dateutil==2.8.2
# 
# pytz==2023.3
# PyYAML==6.0.1
# requests==2.31.0
# six==1.16.0
# SQLAlchemy==2.0.15
# tzdata==2023.3
# uritemplate==4.1.1
# 
# uvicorn==0.22.0

# openpyxl==3.1.2